cleaned, starting other way of handling bounces

+++ model/point/point.cpp
@@ -1,6 +1,8 @@
 #include "point.hpp"
 
-Point::Point(double x, double y) : x{x}, y{y} {}
+#include <cstddef>
+
+Point::Point(size_t x, size_t y) : x{x}, y{y} {}
 
 bool Point::operator==(const Point &other) const {
     return x == other.x and y == other.y;
